fix: organized file structure in frontend
